move router to a more sensible place, rename it to be routers to match django convention of models, forms
